titleシーン固有の機能を分離
