+++ Assets/Models/HashiModel/Hashi_Hashi_BaseColor.png.meta
@@ -147,7 +147,8 @@ TextureImporter:
     secondaryTextures: []
     spriteCustomMetadata:
       entries: []
-    nameFileIdTable: {}
+    nameFileIdTable:
+      Hashi_Hashi_BaseColor_0: -752486226135608161
   mipmapLimitGroupName: 
   pSDRemoveMatte: 0
   userData:
