+++ Assets/Models/OutlineModel.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0f5fddd8e6b124704935991dc4c7f633
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
