+++ Assets/Models/OutlineModel/GoshinbokuOutline.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5fbfdbff36bfd49f389a83b83dd1b088
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
