+++ Assets/Scripts/MenuScene/TitleSceneManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d42a9462695da41219f5a54ea9fb7290
